requirements.txt
# 数据获取和处理
python-binance==1.0.19
pandas==2.0.3
numpy==1.24.3

# 技术指标
ta==0.11.0

# 机器学习
scikit-learn==1.3.0
hmmlearn==0.3.0

# 深度学习
tensorflow==2.13.0
keras==2.13.1

# 任务调度
schedule==1.2.0

# 工具
python-dotenv==1.0.0
tqdm==4.66.1

# 日志和监控
loguru==0.7.2

# 可选：数据可视化
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
